requirements.txt
fastapi[standard]
python-multipart
google-genai
langchain 
langgraph 
langsmith 
google-generativeai 
docker
langchain-google-genai
langfuse
aiohttp
